requirements.txt
opencensus-ext-azure
opencensus-ext-requests
python-dotenv